Separate into Class folders
