+++ src/MemoryDatastore/MemoryDatastore.test.ts
@@ -1,5 +1,5 @@
 import { MemoryDatastore } from './MemoryDatastore'
-import { Datastore, SearchStrategy } from './Datastore'
+import { Datastore, SearchStrategy } from '../Datastore/Datastore'
 
 const readWriteWorks = async (datastore: Datastore): Promise<boolean> => {
   await datastore.write(`key`, `value`)
